requirements.txt
Flask==1.1.4; python_version < '3.0'
Flask==3.0.0; python_version > '3.0'
google-cloud-spanner